+++ xknx/secure/__init__.py
@@ -1,10 +1,10 @@
 """Classes for handling KNX IP Secure."""
-from .keyring import Keyring, load_key_ring
+from .keyring import Keyring, load_keyring
 from .util import bytes_xor, sha256_hash
 
 __all__ = [
     "Keyring",
-    "load_key_ring",
+    "load_keyring",
     "bytes_xor",
     "sha256_hash",
 ]